+++ src/main/java/leets/land/domain/UpDownGame.java
@@ -2,18 +2,9 @@ package leets.land.domain;
 
 import leets.land.domain.status.GuessStatus;
 
-public class UpDownGame {
+public interface UpDownGame {
 
-    private final AnswerNumber answerNumber = new RandomNumberGenerator().generate();
-    private GuessRange guessRange = new GuessRange();
+    GuessStatus play(GuessNumber guessNumber);
 
-    public GuessStatus play(GuessNumber guessNumber) {
-        GuessStatus guessStatus = answerNumber.compare(guessNumber);
-        guessRange = guessStatus.narrowRange(guessNumber, guessRange);
-        return guessStatus;
-    }
-
-    public GuessRange guessRange() {
-        return guessRange;
-    }
+    GuessRange guessRange();
 }
